refactor(manager): split runspec subpackages
