+++ internal/manager/runspec/processor/finalize.go
@@ -1,4 +1,4 @@
-package runspec
+package processor
 
 import (
 	"encoding/json"
