+++ internal/manager/runspec/processor/processor.go
@@ -1,4 +1,4 @@
-package runspec
+package processor
 
 import (
 	"context"
